[REFACTOR/#21] Home 컴포넌트 패키지 구조 수정
