+++ app/src/main/java/com/sopt/shinmungo/presentation/home/HomeScreen.kt
@@ -27,12 +27,12 @@ import com.sopt.shinmungo.R
 import com.sopt.shinmungo.core.designsystem.component.button.RoundedCornerTextButton
 import com.sopt.shinmungo.core.designsystem.theme.ShinMunGoTheme
 import com.sopt.shinmungo.core.state.UiState
-import com.sopt.shinmungo.presentation.home.component.fab.HomeFloatingActionButton
+import com.sopt.shinmungo.presentation.home.component.HomeFloatingActionButton
 import com.sopt.shinmungo.presentation.home.component.group.HomeMonthReportGroup
 import com.sopt.shinmungo.presentation.home.component.group.HomeYearReportGroup
-import com.sopt.shinmungo.presentation.home.component.pager.HomeImageBannerPager
-import com.sopt.shinmungo.presentation.home.component.pager.HomeIssueExamplePager
-import com.sopt.shinmungo.presentation.home.component.topbar.HomeTopBar
+import com.sopt.shinmungo.presentation.home.component.HomeImageBannerPager
+import com.sopt.shinmungo.presentation.home.component.HomeIssueExamplePager
+import com.sopt.shinmungo.presentation.home.component.HomeTopBar
 import com.sopt.shinmungo.presentation.home.state.HomeUiState
 
 @Composable
