+++ app/src/main/java/com/sopt/shinmungo/presentation/home/component/button/BannerIconButton.kt
@@ -6,7 +6,10 @@ import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.sopt.shinmungo.R
 import com.sopt.shinmungo.core.designsystem.theme.ShinMunGoTheme
 import com.sopt.shinmungo.core.extension.noRippleClickable
 
@@ -26,5 +29,14 @@ fun BannerIconButton(
                 shape = RoundedCornerShape(5.dp)
             )
             .noRippleClickable(onClick),
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun BannerIconButtonPreview() {
+    BannerIconButton(
+        imageVector = ImageVector.vectorResource(R.drawable.ic_arrow_right_24),
+        onClick = {}
     )
 }
